XPS: Scale Experimental Data Instead of Simulated Data
Changes the logic for the intensity scaling feature of PR #627
to scale the intensities of the experimental data instead of the
simulated data. This approach is intended to greatly speed up the
response time for the intensity scale widget.

This change also uses `numpy` rather than `pandas` to read the
incoming CSV data, in order to make data manipulation more intuitive.
This change strictly enforces the requirement that the CSV file should
have no header line.